The project "FreeTechGuru" is done on asp.net technology. The objective of this project is to prvide onlne soltion to our regstrd customer.
• The project “FreeTechGuru.” The main objective of this project is to
provide solution for any kind of problem. User should take registration
in our website. Once the user raised the ticket he/she can interact with
the Moderator. The user can set the priority (Priorities have been given
like low, high, medium, urgent and emergency) and status ( i. e. open,
on hold, Insufficient Details, Awaiting for response ,Suspended ,Closed
and Critical ) to their problem so Moderator will solve and respond to
the problem according to the status and priority.  If the problems are
not solved and also if the user is not satisfied with the replies of
Moderator then can send to admin.